SDA-4359 - Refactor BI & send in memory session analytic events
