+++ spec/snippingTool.spec.tsx
@@ -8,7 +8,7 @@ jest.mock('save-svg-as-png', function () {
 
 import { mount, shallow } from 'enzyme';
 import * as React from 'react';
-import { ScreenSnippetActionTypes } from '../src/app/analytics-handler';
+import { ScreenSnippetActionTypes } from '../src/app/bi/analytics-handler';
 import { ScreenShotAnnotation } from '../src/common/ipcEvent';
 import SnippingTool from '../src/renderer/components/snipping-tool';
 import { ipcRenderer } from './__mocks__/electron';
