+++ src/app/bi/analytics-handler.ts
