+++ src/app/main-api-handler.ts
@@ -20,8 +20,8 @@ import { i18n, LocaleType } from '../common/i18n';
 import { logger } from '../common/logger';
 import { whitelistHandler } from '../common/whitelist-handler';
 import { activityDetection } from './activity-detection';
-import { analytics, SDAUserSessionActionTypes } from './analytics-handler';
 import appStateHandler from './app-state-handler';
+import { analytics, SDAUserSessionActionTypes } from './bi/analytics-handler';
 import { closeC9Pipe, connectC9Pipe, writeC9Pipe } from './c9-pipe-handler';
 import { loadC9Shell, terminateC9Shell } from './c9-shell-handler';
 import { getCitrixMediaRedirectionStatus } from './citrix-handler';
